Moved COVID19 Modal
Removed from html.js and put it into its own component, which is loaded in Layout.js after the footer.
